+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# devel
+# 3.11.1.1 (2023-07-26)
 
 * Balance changes are now disabled by default, bug fixes enabled. Will not affect exsting worlds/configs.
 * Modularized Portal block configuration; you can now define your own blocks to build Vault Portals out of.